ml_project/requirements.txt
brotli==1.0.9
brotlipy==0.7.0
click==8.0.4
ConfigParser==5.3.0
cryptography==38.0.1
Cython==0.29.32
dl==0.1.0
Faker==15.1.1
HTMLParser==0.0.2
hydra-core==1.2.0
importlib_metadata==4.11.3
ipaddr==2.2.0
keyring==23.4.0
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.1
matplotlib==3.5.2
mypy_extensions==0.4.3
omegaconf==2.2.3
ordereddict==1.1
pandas==1.4.4
protobuf==4.21.9
pyOpenSSL==22.1.0
python-dotenv==0.21.0
pywin32==302
scikit_learn==1.1.3
seaborn==0.11.1
sweetviz==2.1.4
wincertstore
zipp==3.8.0
